feat: first-and-last-pages' chunk selector
For the extraction of some fields, a selection of the chunks in the
first and last pages has been implemented

NB: we already guess that some fields will not be always extractable
from either the first or last pages, e.g. the names behind an
intervention can be in a signature at the end of a document, but not in
the last pages because these pages are appendix
The best solution will be to couple embedding comparison and named
entities extraction.