fix: valid var name for dspy module output field
With this, the table displayed by the evaluation results
is well rendered.
This table is not suitable for user-friendly monitoring, as the number
of fields is big and the columns are not paired by dspy according to the
predicted and in-example fields
However, we are sure the processing by dspy (and perhaps mlflow) is
standardized.

TODO: finish the processing of the result dataframe